pyproject.toml
[project]
name = "milvus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chonkie[semantic]>=0.2.2",
    "crewai>=0.102.0",
    "fastapi>=0.115.8",
    "numpy>=2.2.3",
    "openai>=1.63.2",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "pymilvus>=2.5.4",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
